purgeDatabase waits mongodb connection

+++ lib/util/test.coffee
@@ -4,4 +4,10 @@ mongoose = require 'mongoose'
 module.exports =
 
   purgeDatabase: (callback) ->
-    mongoose.connection.db.dropDatabase callback
+    # db 以下の操作は mongoose 側の接続管理が適用されないのでリトライをする
+    setTimeout ->
+      if mongoose.connection.readyState is 1
+        mongoose.connection.db.dropDatabase callback
+      else
+        setTimeout arguments.callee, 50
+    , 1
